Moved App into its own dir inside src and updated imports
