+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './App/App';
 import reportWebVitals from './reportWebVitals';
 
 import './normalize/normalize.scss'
